update to 1.7.0

+++ unittests/special_accounts_tests.cpp
@@ -3,23 +3,21 @@
  *  @copyright defined in eos/LICENSE
  */
 #include <algorithm>
-#include <vector>
 #include <iterator>
-#include <boost/test/unit_test.hpp>
+#include <vector>
 
 #include <eosio/chain/controller.hpp>
 #include <eosio/chain/exceptions.hpp>
 #include <eosio/chain/permission_object.hpp>
 #include <eosio/chain/global_property_object.hpp>
-
 #include <eosio/testing/tester.hpp>
 
 #include <fc/crypto/digest.hpp>
 
-#include <boost/test/unit_test.hpp>
 #include <boost/range/algorithm/find.hpp>
 #include <boost/range/algorithm/find_if.hpp>
 #include <boost/range/algorithm/permutation.hpp>
+#include <boost/test/unit_test.hpp>
 
 using namespace eosio;
 using namespace chain;
@@ -38,14 +36,14 @@ BOOST_FIXTURE_TEST_CASE(accounts_exists, tester)
       auto nobody = chain1_db.find<account_object, by_name>(config::null_account_name);
       BOOST_CHECK(nobody != nullptr);
       const auto& nobody_active_authority = chain1_db.get<permission_object, by_owner>(boost::make_tuple(config::null_account_name, config::active_name));
-      BOOST_CHECK_EQUAL(nobody_active_authority.auth.threshold, 1);
-      BOOST_CHECK_EQUAL(nobody_active_authority.auth.accounts.size(), 0);
-      BOOST_CHECK_EQUAL(nobody_active_authority.auth.keys.size(), 0);
+      BOOST_CHECK_EQUAL(nobody_active_authority.auth.threshold, 1u);
+      BOOST_CHECK_EQUAL(nobody_active_authority.auth.accounts.size(), 0u);
+      BOOST_CHECK_EQUAL(nobody_active_authority.auth.keys.size(), 0u);
 
       const auto& nobody_owner_authority = chain1_db.get<permission_object, by_owner>(boost::make_tuple(config::null_account_name, config::owner_name));
-      BOOST_CHECK_EQUAL(nobody_owner_authority.auth.threshold, 1);
-      BOOST_CHECK_EQUAL(nobody_owner_authority.auth.accounts.size(), 0);
-      BOOST_CHECK_EQUAL(nobody_owner_authority.auth.keys.size(), 0);
+      BOOST_CHECK_EQUAL(nobody_owner_authority.auth.threshold, 1u);
+      BOOST_CHECK_EQUAL(nobody_owner_authority.auth.accounts.size(), 0u);
+      BOOST_CHECK_EQUAL(nobody_owner_authority.auth.keys.size(), 0u);
 
       auto producers = chain1_db.find<account_object, by_name>(config::producers_account_name);
       BOOST_CHECK(producers != nullptr);
@@ -56,7 +54,7 @@ BOOST_FIXTURE_TEST_CASE(accounts_exists, tester)
       auto expected_threshold = (active_producers.producers.size() * 2)/3 + 1;
       BOOST_CHECK_EQUAL(producers_active_authority.auth.threshold, expected_threshold);
       BOOST_CHECK_EQUAL(producers_active_authority.auth.accounts.size(), active_producers.producers.size());
-      BOOST_CHECK_EQUAL(producers_active_authority.auth.keys.size(), 0);
+      BOOST_CHECK_EQUAL(producers_active_authority.auth.keys.size(), 0u);
 
       std::vector<account_name> active_auth;
       for(auto& apw : producers_active_authority.auth.accounts) {
@@ -64,7 +62,7 @@ BOOST_FIXTURE_TEST_CASE(accounts_exists, tester)
       }
 
       std::vector<account_name> diff;
-      for (int i = 0; i < std::max(active_auth.size(), active_producers.producers.size()); ++i) {
+      for (size_t i = 0; i < std::max(active_auth.size(), active_producers.producers.size()); ++i) {
          account_name n1 = i < active_auth.size() ? active_auth[i] : (account_name)0;
          account_name n2 = i < active_producers.producers.size() ? active_producers.producers[i].producer_name : (account_name)0;
          if (n1 != n2) diff.push_back((uint64_t)n2 - (uint64_t)n1);
